#2244 Refactor obsolete WebBasedTestCase into a static test utility to better reflect its purpose
